rapidresponse_lens/requirements.txt
Adafruit-Blinka==8.6.1
adafruit-circuitpython-busdevice==5.2.3
adafruit-circuitpython-neopixel==6.3.6
adafruit-circuitpython-pixelbuf==1.1.8
adafruit-circuitpython-requests==1.12.10
adafruit-circuitpython-typing==1.8.2
Adafruit-PlatformDetect==3.33.0
Adafruit-PureIO==1.1.9
asyncio==3.4.3
azure-core==1.26.0
azure-cosmos==4.3.0
azure-iot-device==2.12.0
azure-storage-blob==12.14.1
blinker==1.4
blinkt==0.1.2
blobconverter==1.2.9
depthai==2.18.0.0
-e git+https://github.com/luxonis/depthai.git@e377e9bcf7caa87f421fd95e53d8263d23bec764#egg=depthai_sdk&subdirectory=depthai_sdk
distlib==0.3.5
distro==1.5.0
dlib==19.24.0
docutils==0.16
ffmpy3==0.2.4
filelock==3.7.1
Flask==1.1.2
Flask-Caching==2.0.1
imagezmq==1.1.1
imutils==0.5.4
ipykernel==5.4.3
ipython==7.20.0
ipython_genutils==0.2.0
Jinja2==2.11.3
matplotlib==3.3.4
numpy==1.21.6
oauthlib==3.1.0
opencv-contrib-python==4.4.0.46
opencv-python==4.5.1.48
paho-mqtt==1.6.1
pantilthat==0.0.7
picamera==1.13
Pillow==9.0.1
pylint==2.7.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyserial==3.5b0
PyTurboJPEG==1.6.4
pyusb==1.2.1
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.0.0
requests-unixsocket==0.3.0
responses==0.12.1
rpi-ws281x==4.3.4
RPi.GPIO==0.7.0
scipy==1.6.0
simplejson==3.17.2
spidev==3.5
ssh-import-id==5.10
urllib3==1.26.5
virtualenv==20.16.3
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
Werkzeug==1.0.1